requirements.txt
flask==2.3.3
flask-wtf==1.2.1
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.23
Werkzeug==2.3.7
openai==1.6.1
anthropic==0.8.1
google-generativeai==0.3.2
notion-client==2.0.0 